Update docker-compose.yml
Added redis server for rate limiting and put its port and host as environments in the api server. I was getting errors that version 3.7 was unsupported by my docker-toolbox so I changed it to version 3.3. Then had to remove the name under the mysql-data volume at the bottom as it wasn't compatible with version 3.3.